# Backend/MivoraSwagger.yaml
openapi: 3.0.4
info:
  title: Mivora - 2025
  description: |-
    This OpenAPI spec describes the backend API for Mivora, a platform that handles users, events, tickets, and real-time event messaging. It includes full user account flows such as registration, login, logout, token refresh, profile updates, avatar uploads, email verification, and password-reset procedures. Event organizers can create, update, publish, cancel, and manage event posters, while both organizers and attendees can send and fetch event messages. Users can browse published events, view details, and book tickets. The ticketing system supports booking, listing with filters, retrieving individual tickets, canceling or checking in tickets, and QR-code-based scanning. All sensitive and user-specific operations use JWT bearer authentication, and list endpoints follow a consistent pagination pattern.
    some useful links:
    - [The Mivora repository](https://github.com/SantiagoDeStefano/Mivora)
    - [The source API definition for Mivora](https://github.com/SantiagoDeStefano/Mivora/blob/main/Backend/MivoraSwagger.yaml)

  version: 0.0.1
servers:
  - url: /
    description: Same host as Swagger
  - url: http://47.129.118.52:4000
    description: EC2 server
  - url: http://localhost:4000
    description: Local server
  - url: http://26.35.82.76:4000
    description: LAN server
  - url: https://khoinguyenpham.name.vn
    description: Production api server
...
paths:
...
  /api/v1/events:
    get:
      tags:
        - events
      summary: Get all published events
      operationId: getAllPublishedEvents
      parameters:
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            default: 20
          description: Number of items per page
        - name: page
          in: query
          required: true
          schema:
            type: integer
            default: 1
          description: Page number
        - name: q
          in: query
          required: false
          example: Syn
          schema:
            type: string
          description: Search query for event title
      responses:
        '200':
          description: Get events successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get events successfully
                  result:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                              example: 3b1bf6c6-4fa0-4026-95a2-be76908d1f5d
                            organizer_name:
                              type: string
                              example: Lam Uyen Phuong
                            title:
                              type: string
                              example: MEOMEOMEO
                            description:
                              type: string
                              example: tại sao lá cây lại có màu xanh
                            poster_url:
                              type: string
                              format: uri
                              example: https://example.com/poster.jpg
                            location_text:
                              type: string
                              example: 123 Main Street, San Francisco, CA
                            start_at:
                              type: string
                              format: date-time
                              example: 2025-12-06T14:03:00.000Z
                            end_at:
                              type: string
                              format: date-time
                              example: 2025-12-13T14:03:00.000Z
                            price_cents:
                              type: integer
                              example: 34300
                            checked_in:
                              type: integer
                              example: 0
                            capacity:
                              type: integer
                              example: 3434
                            status:
                              type: string
                              example: published
                            total_count:
                              type: string
                              example: '158'
                      limit:
                        type: integer
                        example: 20
                      page:
                        type: integer
                        example: 1
                      total_page:
                        type: integer
                        example: 8
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Validation error
                  errors:
                    type: object
                    properties:
                      q:
                        type: object
                        properties:
                          type:
                            type: string
                            example: field
                          value:
                            type: string
                            example: '0'
                          msg:
                            type: string
                            example: Search length must be between 3 and 20
                          path:
                            type: string
                            example: q
                          location:
                            type: string
                            example: query
    post:
      tags:
        - events
      summary: Create a new event
      operationId: createEvent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Tech Innovation Summit 2027
                description:
                  type: string
                  example: A conference exploring emerging technologies and innovation trends.
                location_text:
                  type: string
                  example: 123 Main Street, San Francisco, CA
                start_at:
                  type: string
                  format: date-time
                  example: 2025-12-06T11:49:27.918Z
                end_at:
                  type: string
                  format: date-time
                  example: 2025-12-06T11:49:28.918Z
                price_cents:
                  type: integer
                  example: 2500
                capacity:
                  type: integer
                  example: 150
      responses:
        '200':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event created successfully
                  result:
                    $ref: '#/components/schemas/Event'

        '403':
          description: Must be organizer
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Must be organizer'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Validation error
                  errors:
                    type: object
                    properties:
                      price_cents:
                        type: object
                        properties:
                          type:
                            type: string
                            example: field
                          msg:
                            type: string
                            example: Event's price is required
                          path:
                            type: string
                            example: price_cents
                          location:
                            type: string
                            example: body
...
  /api/v1/events/{event_id}/messages:
    post:
      tags:
        - events
      summary: Send message to event attendees
      operationId: sendMessageToEventAttendees
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the event to send message
          example: 084bc4e5-8fe0-4394-a85d-1e65ed3dae16
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: Xin chao mng
      responses:
        '200':
          description: Create event message successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get event messages successfully
                  result:
                    $ref: '#/components/schemas/EventMessage'

        '403':
          description: Current user is not event's attendee or organizer
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Current user is not event's attendee or organizer"
    get:
      tags:
        - events
      summary: Get event messages
      operationId: getEventMessages
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the event to get messages
          example: 084bc4e5-8fe0-4394-a85d-1e65ed3dae16
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            default: 20
          description: Number of items per page
        - name: page
          in: query
          required: true
          schema:
            type: integer
            default: 1
          description: Page number
      responses:
        '200':
          description: Get event messages successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get event messages successfully
                  result:
                    type: object
                    properties:
                      messages:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventMessage'
                      limit:
                        type: integer
                        example: 20
                      page:
                        type: integer
                        example: 1
                      total_page:
                        type: integer
                        example: 3

        '403':
          description: Current user is not event's attendee or organizer
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Current user is not event's attendee or organizer"
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 6a2f69f9-321b-4c73-adc5-eba50c761bdc
        title:
          type: string
          example: Tech Innovation Summit 2027
        description:
          type: string
          example: A conference exploring emerging technologies and innovation trends.
        poster_url:
          type: string
          format: uri
          example: https://mivora-ap-southeast-1.s3.ap-southeast-1.amazonaws.com/avatar-images/gfu4o33ca43nvcwyv4vay18m1.jpg
        location_text:
          type: string
          example: 123 Main Street, San Francisco, CA
        start_at:
          type: string
          format: date-time
          example: 2025-12-06T11:49:27.918Z
        end_at:
          type: string
          format: date-time
          example: 2025-12-06T11:49:28.918Z
        price_cents:
          type: integer
          example: 2500
        checked_in:
          type: integer
          example: 0
        capacity:
          type: integer
          example: 150
        status:
          type: string
          example: draft
...
    EventMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        user_avatar_url:
          type: string
          format: uri
      required:
        - id
        - content
        - created_at
        - user_id
        - user_name
        - user_avatar_url
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Use the `Bearer` scheme to authenticate requests. Include the token in the `Authorization` header as follows:
        ```
        Authorization: Bearer <your_token>
        ```